# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API Documentation",
        "description": "API documentation for Point of Sale application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get sales reports for all periods",
                "operationId": "81a8afd2aa3c69c6b5565682c3a78dd7",
                "responses": {
                    "200": {
                        "description": "Report data retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/sales-history": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get paginated sales history",
                "operationId": "20d9dbeffb4e2b61043b05b6f2babad0",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "semua"
                        },
                        "example": "hari_ini"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        },
                        "example": 10
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "mie"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales history retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mie Goreng Aceh"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 3500
                    },
                    "description": {
                        "type": "string",
                        "example": "Makanan khas Aceh"
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "stocks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Stock"
                        }
                    }
                },
                "type": "object"
            },
            "Stock": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "stock_on_hand": {
                        "type": "integer",
                        "example": 40
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Makanan"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "trx_type": {
                        "type": "string",
                        "enum": [
                            "SALE",
                            "PURCHASE",
                            "ADJUSTMENT"
                        ]
                    },
                    "trx_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-03"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "CASH"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 7000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransactionItem"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "TransactionItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transaction_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 3500
                    },
                    "line_price": {
                        "type": "number",
                        "format": "float",
                        "example": 7000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
...
            "ReportData": {
                "properties": {
                    "total_pendapatan": {
                        "type": "number",
                        "format": "float"
                    },
                    "pendapatan_vs_sebelumnya": {
                        "properties": {
                            "nilai_sebelumnya": {
                                "type": "number",
                                "format": "float"
                            },
                            "persentase_perubahan": {
                                "type": "number",
                                "format": "float"
                            }
                        },
                        "type": "object"
                    },
                    "total_transaksi": {
                        "type": "integer"
                    },
                    "rata_rata_keranjang": {
                        "type": "number",
                        "format": "float"
                    },
                    "tren_penjualan": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "total": {
                                    "type": "number",
                                    "format": "float"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "produk_terlaris": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "total_quantity": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "transaksi_terakhir": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    }
                },
                "type": "object"
            },
            "ReportResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "hari_ini": {
                                "$ref": "#/components/schemas/ReportData"
                            },
                            "minggu_ini": {
                                "$ref": "#/components/schemas/ReportData"
                            },
                            "bulan_ini": {
                                "$ref": "#/components/schemas/ReportData"
                            },
                            "tahun_ini": {
                                "$ref": "#/components/schemas/ReportData"
                            },
                            "sepanjang_masa": {
                                "$ref": "#/components/schemas/ReportData"
                            },
                            "grafik_data": {
                                "properties": {
                                    "minggu_ini": {
                                        "properties": {
                                            "labels": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "values": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "bulan_ini": {
                                        "properties": {
                                            "labels": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "values": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "tahun_ini": {
                                        "properties": {
                                            "labels": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "values": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}